Arreglando

+++ app/src/main/java/com/example/mapptuu/data/model/Activity.kt
@@ -12,5 +12,5 @@ data class Activity(
     val longitude:String,
     val name:String,
     val ownerId:String,
-    val rating:Int
+    val rating:Float
 )
